+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to the "emacs-mcx" extension will be documented in this file.
 
 ## [Unreleased]
+
+### Add
+- `emacs-mcx.enableDigitArgument` config, #1256.
+
 ## [0.40.1] - 2022-03-20
 ### Fix
 - Release process, #1254.