+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 
 ## [Unreleased]
 ### Add
+- `emacs-mcx.enableDigitArgument` config, #1256.
 - Regexp searches, #1257.
 
 ## [0.40.1] - 2022-03-20